requirements.txt
flask
numpy==2.1.3
scikit-learn
scipy 
matplotlib
torch
open-clip-torch
pandas